% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countpackdiff.R
\name{countpackdiff}
\alias{countpackdiff}
\title{Count number of packages out of sync}
\usage{
countpackdiff(workdir = getwd())
}
\arguments{
\item{workdir}{(optional): an alternative path to the package list files}
}
\value{
Number of packages out of sync
}
\description{
Counts the number of packages out of sync between two list of packages saved
with \emph{savepackages} in your working directory
}
