% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdropboxdir.R
\name{getdropboxdir}
\alias{getdropboxdir}
\title{Returns the Dropbox directory}
\usage{
getdropboxdir()
}
\arguments{
\item{None}{}
}
\value{
A filepath string
}
\description{
Finds the current Dropbox directory on your system and returns it as a file-path
}
\note{
This is just using the paths described in the Dropbox documentation, after determining the system name to be one of "Windows", "Darwin" or "Linux"
}
\references{
\url{https://www.dropbox.com/help/desktop-web/find-folder-paths}
}
