% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagelist.R
\name{PkgsToInstall}
\alias{PkgsToInstall}
\title{Gives the text to install packages}
\usage{
PkgsToInstall()
}
\description{
`pkgs_list()` is an
  [addin](http://rstudio.github.io/rstudioaddins/) that creates you code
  to install the selected packages from existing code. Appears as "Reprex selection"
  in the RStudio Addins menu. Heavy users might want to [create a keyboard
  shortcut](https://support.rstudio.com/hc/en-us/articles/206382178-Customizing-Keyboard-Shortcuts).
}
