% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagelist.R
\name{rstudio_context}
\alias{rstudio_context}
\title{Helper Functions blatantly copied from reprex package}
\usage{
rstudio_context()
}
\description{
Following functions are used to get text from the current selection
in your Rstudio
}
