% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagelist.R
\name{rstudio_selection}
\alias{rstudio_selection}
\title{Helper Functions blatantly copied from reprex package}
\usage{
rstudio_selection(context = rstudio_context())
}
\arguments{
\item{context}{extract Rstudio context}
}
\description{
Following functions are used to get text from the current selection
in your Rstudio
}
