% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savepackages.R
\name{savepackages}
\alias{savepackages}
\title{Save package list}
\usage{
savepackages(workdir = getwd())
}
\arguments{
\item{workdir}{(optional): an alternative path to save to}
}
\value{
none
}
\description{
Saves all currently installed packages to an RDS file in your working directory
}
