% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncpackages.R
\name{syncpackages}
\alias{syncpackages}
\title{Save package list}
\usage{
syncpackages(workdir = getwd())
}
\arguments{
\item{workdir}{(optional): an alternative path to the package list file}
}
\value{
none
}
\description{
Installs all packages from a list saved with \emph{savepackages} in your working directory
}
