% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagelist.R
\name{wrap_it_up}
\alias{wrap_it_up}
\title{To wrap up the selected text with install.packages()}
\usage{
wrap_it_up(text = "library('tidyverse')")
}
\arguments{
\item{text}{the selected text from RStudio - default is library(tidyverse)}
}
\description{
Selected Text from RStudio (Current Selection) is reformatted to
include install.packages
}
